requirements.txt
awsebcli==3.7.7
blessed==1.9.5
boto==2.42.0
botocore==1.4.43
cement==2.8.2
colorama==0.3.7
defusedxml==0.4.1
Django==1.10.1
django-allauth==0.27.0
django-ses==0.8.1
django-storages==1.5.0
django-widget-tweaks==1.4.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
oauthlib==2.0.0
pathspec==0.3.4
psycopg2==2.6.2
python-dateutil==2.5.3
python3-openid==3.0.10
PyYAML==3.11
requests==2.11.1
requests-oauthlib==0.6.2
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.7
websocket-client==0.37.0
